LRCI-397 Shouldn't be an inner class anymore
